// frontend/src/generated-client/api.ts
export interface Company {
    /**
     * 
     * @type {number}
     * @memberof Company
     */
    'companyID'?: number;
    /**
     * 
     * @type {string}
     * @memberof Company
     */
    'companyName'?: string | null;
}
// ...
export class CompaniesApi extends BaseAPI {
    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof CompaniesApi
     */
    public companiesGet(options?: AxiosRequestConfig) {
        return CompaniesApiFp(this.configuration).companiesGet(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof CompaniesApi
     */
    public companiesIdDelete(id: number, options?: AxiosRequestConfig) {
        return CompaniesApiFp(this.configuration).companiesIdDelete(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof CompaniesApi
     */
    public companiesIdGet(id: number, options?: AxiosRequestConfig) {
        return CompaniesApiFp(this.configuration).companiesIdGet(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} id 
     * @param {Company} [company] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof CompaniesApi
     */
    public companiesIdPut(id: number, company?: Company, options?: AxiosRequestConfig) {
        return CompaniesApiFp(this.configuration).companiesIdPut(id, company, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {Company} [company] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof CompaniesApi
     */
    public companiesPost(company?: Company, options?: AxiosRequestConfig) {
        return CompaniesApiFp(this.configuration).companiesPost(company, options).then((request) => request(this.axios, this.basePath));
    }
}